requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
bcrypt==4.2.0
stripe==10.12.0
pydantic[email]==2.9.0
psycopg2-binary==2.9.9
